refactor: split collections constants
