+++ src/console/src/memory.rs
@@ -3,7 +3,7 @@ use crate::types::state::{StableState, State};
 use ic_stable_structures::memory_manager::{MemoryId, MemoryManager};
 use ic_stable_structures::DefaultMemoryImpl;
 use ic_stable_structures::StableBTreeMap;
-use junobuild_collections::constants::DEFAULT_ASSETS_COLLECTIONS;
+use junobuild_collections::constants_assets::DEFAULT_ASSETS_COLLECTIONS;
 use junobuild_shared::types::memory::Memory;
 use junobuild_storage::types::state::StorageHeapState;
 use std::cell::RefCell;
