+++ src/libs/satellite/src/db/impls.rs
@@ -4,7 +4,7 @@ use candid::Principal;
 use ic_cdk::api::time;
 use ic_stable_structures::storable::Bound;
 use ic_stable_structures::Storable;
-use junobuild_collections::constants::DEFAULT_DB_COLLECTIONS;
+use junobuild_collections::constants_db::DEFAULT_DB_COLLECTIONS;
 use junobuild_collections::types::rules::{Memory, Rule};
 use junobuild_shared::serializers::{deserialize_from_bytes, serialize_to_bytes};
 use junobuild_shared::types::state::{Timestamp, UserId, Version};
