+++ src/libs/storage/src/assert.rs
@@ -8,7 +8,8 @@ use crate::types::state::FullPath;
 use crate::types::store::{Asset, AssetAssertUpload, Batch};
 use candid::Principal;
 use junobuild_collections::assert_stores::{assert_create_permission, assert_permission};
-use junobuild_collections::constants::{DEFAULT_ASSETS_COLLECTIONS, SYS_COLLECTION_PREFIX};
+use junobuild_collections::constants::SYS_COLLECTION_PREFIX;
+use junobuild_collections::constants_assets::DEFAULT_ASSETS_COLLECTIONS;
 use junobuild_collections::types::core::CollectionKey;
 use junobuild_collections::types::rules::Rule;
 use junobuild_shared::assert::{assert_description_length, assert_max_memory_size};
